requirements.txt
APScheduler==3.6.3
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
sqlalchemy-mixins==1.2.1
tzlocal==2.1
psycopg2==2.8.5
python-decouple==3.3APScheduler==3.6.3
boto3==1.16.4
botocore==1.19.4
jmespath==0.10.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
sqlalchemy-mixins==1.2.1
sqlparse==0.4.1
tabulate==0.8.7
tzlocal==2.1
urllib3==1.25.11
yoyo-migrations==7.2.0
